color, backgroungColor componentをon offできるようにしました
