+++ src/components/cssEditArea/CssEditArea.tsx
@@ -1,8 +1,7 @@
 import React from 'react'
 import { Box, Button, Flex, Text } from '@chakra-ui/react'
 import { AddIcon } from '@chakra-ui/icons'
-import { BackgroundColor } from './backgroundColor/BackgroundColor'
-import { Color } from './color/color'
+import { EditColor } from './color/EditColor'
 import { Padding } from './padding/Padding'
 import { useAppDispatch, useAppSelector } from '../../hooks'
 import { setIsOpen } from '../cssCustomArea/cssCustomAreaSlice'
@@ -14,6 +13,7 @@ import { EditBorderStyle } from './borderStyle/EditBorderStyle'
 import { EditBorderRadius } from './borderRadius/EditBorderRadius'
 import { EditWidth } from './width/EditWidth'
 import { EditHeight } from './height/EditHeight'
+import { EditBackgroundColor } from './backgroundColor/EditBackgroundColor'
 
 export const CssEditArea = () => {
     const dispatch = useAppDispatch()
@@ -26,8 +26,8 @@ export const CssEditArea = () => {
             </Text>
             <EditWidth />
             <EditHeight />
-            <Color />
-            <BackgroundColor />
+            <EditColor />
+            <EditBackgroundColor />
             <Padding />
             <EditBorderColor />
             <EditBorderStyle />
